+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ### 0.2 — Unreleased
 
-- Added `Lightstreamer::Session#disconnect` and `Lightstreamer::StreamConnection#disconnect`
+- Added `Lightstreamer::Session#subscribed?` and `Lightstreamer::Session#disconnect`
 - Fixed `Lightstreamer::ProtocolError#code` not being set
-- Renamed the `--address` option to the command-line client to `--server-url`
+- Renamed the command-line client's `--address` option to `--server-url`
 
 ### 0.1 — July 22, 2016